change field average_coach_rating to Float

+++ src/schema.js
@@ -27,7 +27,7 @@ const typeDefs = gql`
 		chatActive: Boolean
 		reviewsRecieved: [Review!]
 		reviewsGiven: [Review!]
-		average_coach_rating(microservice: String): Int
+		average_coach_rating(microservice: String): Float
 	}
 
 	type Review @key(fields: "id") @key(fields:"job"){
